refactor: reorganize imports and remove unused flow files in chat, portfolio analysis, and RAG examples
